highScore Algorithm completed and tested - Everything is a double though, and I would prefer them to be an int.  Remember to ask Brandon about this.

+++ solarTrail/src/buyi/cit260/solarTrail/control/GameControl.java
@@ -23,7 +23,7 @@ public class GameControl {
         return resourcesLeft;
     }
         public double highScore(double difficultyLvl, double crewAlive, double eventsCleared, 
-            int artifactScore) {
+            double artifactScore) {
         if (difficultyLvl > 3)
             return -1;
         if (crewAlive < 0)
